+++ frontend/src/views/FileManager/ExplorerUI/file-icons.ts
@@ -2,25 +2,29 @@ import type { IEntry } from '@/types/server'
 import {
   regArchiveFormat,
   regAudioFormat,
+  regExcelFormat,
   regImageFormat,
+  regPdfFormat,
+  regPowerpointFormat,
   regSupportedTextFormat,
   regVideoFormat,
-  regWindowsExeFormat,
+  regWindowsFormat,
+  regWordFormat,
 } from '@/utils/is'
 
 export function getFileIconClass(item?: IEntry) {
   if (!item) {
-    return 'mdi-file-question'
+    return 'mdi-file-question-outline'
   }
   const { isDirectory, ext } = item
   if (isDirectory) {
-    return 'mdi-folder'
+    return 'mdi-folder-outline'
   }
   if (regArchiveFormat.test(ext)) {
     return `mdi-zip-box`
   }
   if (regImageFormat.test(ext)) {
-    return `mdi-file-image`
+    return `mdi-file-image-outline`
   }
   if (regAudioFormat.test(ext)) {
     return `mdi-music-circle-outline`
@@ -28,8 +32,20 @@ export function getFileIconClass(item?: IEntry) {
   if (regVideoFormat.test(ext)) {
     return `mdi-file-video-outline`
   }
-  if (regWindowsExeFormat.test(ext)) {
-    return `mdi-microsoft-windows`
+  if (regWindowsFormat.test(ext)) {
+    return `mdi-microsoft`
+  }
+  if (regPdfFormat.test(ext)) {
+    return `mdi-file-pdf-outline`
+  }
+  if (regExcelFormat.test(ext)) {
+    return `mdi-file-excel-outline`
+  }
+  if (regWordFormat.test(ext)) {
+    return `mdi-file-word-outline`
+  }
+  if (regPowerpointFormat.test(ext)) {
+    return `mdi-file-powerpoint-outline`
   }
   if (regSupportedTextFormat.test(ext)) {
     if (/.vue$/.test(ext)) {
@@ -53,7 +69,7 @@ export function getFileIconClass(item?: IEntry) {
     if (/.md$/.test(ext)) {
       return `mdi-language-markdown-outline`
     }
-    return `mdi-text-box-outline`
+    return `mdi-file-document-outline`
   }
-  return 'mdi-file-question'
+  return 'mdi-file-question-outline'
 }